requirements.txt
annotated-types==0.7.0
anyio==4.4.0
asgiref==3.8.1
beautifulsoup4==4.12.3
build==1.2.1
CacheControl==0.14.0
certifi==2024.7.4
charset-normalizer==3.3.2
cleo==2.1.0
clickhouse-connect==0.7.15
colorama==0.4.6
crashtest==0.4.1
Cython==3.0.10
distlib==0.3.8
distro==1.9.0
Django==5.0.6
django-cors-headers==4.4.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
dulwich==0.21.7
fastjsonschema==2.20.0
filelock==3.15.4
gunicorn==22.0.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
importlib_metadata==8.0.0
installer==0.7.0
jaraco.classes==3.4.0
keyring==24.3.1
lz4==4.3.3
more-itertools==10.3.0
msgpack==1.0.8
numpy==2.0.0
openai==1.35.10
packaging==24.1
pandas==2.2.2
pexpect==4.9.0
pkginfo==1.11.1
platformdirs==4.2.2
poetry==1.8.3
poetry-core==1.9.0
poetry-plugin-export==1.8.0
ptyprocess==0.7.0
py4j==0.10.9.7
pydantic==2.8.2
pydantic_core==2.20.1
PyJWT==2.8.0
pyproject_hooks==1.1.0
pyspark==3.5.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
pywin32-ctypes==0.2.2
rapidfuzz==3.9.4
requests==2.32.3
requests-toolbelt==1.0.0
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
soupsieve==2.5
sqlparse==0.5.0
tomlkit==0.12.5
tqdm==4.66.4
trove-classifiers==2024.7.2
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
virtualenv==20.26.3
zipp==3.19.2
zstandard==0.22.0
